inv all + monitor: share one check list, add ifil/glf, drop dead code
`inv all` and `inv monitor` each hard-coded the same six-check list,
and the comment in monitor.go acknowledged the drift risk. Adding a
new check meant editing both places. This extracts the list into
cmd/invariants/checks.go::Checks(epoch, tolerance) and has both
surfaces consume it — one source of truth, one place to edit when a
new invariant joins the aggregate gate.

While doing the refactor, also wires in the two invariants that had
sat in all.go's TODO since v2: `pool ifil --per-depositor` and
`pool glf --per-holder`. Both already exit non-zero on failure via
log.Fatalf("FAIL: ..."), so they were always aggregate-suitable —
just missing from the list. With the iFIL backfill from the dump-
restore work last week, these are the two checks that proved most
useful for catching daemon-era drift; running them on the monitor's
1h tick is cheap insurance.

Cleanup of code that no longer matches the protocol's current state
(we repurposed this repo from a stale snapshot to the live source of
truth — leftover artifacts here are confusing, not safe):

  cmd/invariants/agentecon.go             entire file — `agent econ`
                                          subcommand was deprecated and
                                          never worked end-to-end against
                                          the new schema. Compared API
                                          principal vs node principal but
                                          the API endpoint it hit doesn't
                                          surface the latter at the right
                                          shape anymore.
  agents.go::AgentEconJSON                only consumer was agentecon.go
  agents.go::AgentEconResult              same
  agents.go::GetAgentEconFromAPI          same
  agents.go::GetAgentEconFromNode         same
  metrics.go::GetMetricsFromAPI           non-height variant; only caller
                                          was metrics_test.go.
  metrics_test.go                         test of the above only-test-
                                          consumer-of-dead-code dead
                                          function. The test's contract
                                          (API totalAssets == chain
                                          totalAssets at the API's
                                          reported height) is now covered
                                          by `inv pool metrics`, which
                                          decomposes per-component and
                                          tolerance-checks instead of
                                          equality-asserting.

Build clean. Two pre-existing vet warnings in minerliquidation.go are
unrelated (printf format mismatches); left for a separate cleanup.
